Fixes the go race tests.
We used these iterators with no mutex, back when iterator performance was
one of our main blockers, and this step was thought to be of heavy performance impact.

(It does add a notable delay to iteration steps, but it was later learned
the bulk of the problem was due to IAVL & LevelDB interaction)

Its overdue to have turned this off just for simplicity of tests passing.
We still should update the BTree library and remove the need for this
multiple process terrible hack that exists now, but thats work for the future.
(There is no race condition if you look into the details of whats happening,
the gorace detector is off here!)